added a new home screen file, changed class names to have the same format, renamed files to have the same format
